+++ Joker.Kafka/Extensions/KSql/Query/Record.cs
@@ -1,0 +1,7 @@
+﻿namespace Kafka.DotNet.ksqlDB.Extensions.KSql.Query
+{
+  public class Record
+  {
+    public long RowTime { get; set; }
+  }
+}