+++ Samples/Kafka/Kafka.DotNet.ksqlDB.Sample/Model/Tweet.cs
@@ -1,8 +1,9 @@
 ﻿using System.Text.Json.Serialization;
+using Kafka.DotNet.ksqlDB.Extensions.KSql.Query;
 
 namespace Kafka.DotNet.ksqlDB.Sample.Model
 {
-  public class Tweet
+  public class Tweet : Record
   {
     public int Id { get; set; }