+++ Samples/Kafka/Kafka.DotNet.ksqlDB.Sample/Program.cs
@@ -5,6 +5,7 @@ using Kafka.DotNet.ksqlDB.Extensions.KSql.Linq;
 using Kafka.DotNet.ksqlDB.Extensions.KSql.Query;
 using Kafka.DotNet.ksqlDB.Sample.Model;
 using Kafka.DotNet.ksqlDB.Sample.Observers;
+using Kafka.DotNet.ksqlDB.Sample.QueryStreams;
 
 namespace Kafka.DotNet.ksqlDB.Sample
 {
@@ -15,17 +16,18 @@ namespace Kafka.DotNet.ksqlDB.Sample
       Console.WriteLine("Subscription started");
 
       var ksqlDbUrl = @"http:\\localhost:8088";
-
+      ToQueryStringExample(ksqlDbUrl);
       using var disposable = new KQueryStreamSet<Tweet>(new QbservableProvider(ksqlDbUrl))
         .Where(p => p.Message != "Hello world" || p.Id == 1)
-        .Select(l => new { l.Message, l.Id })
+        //.Where(p => p.RowTime >= 1510923225000) //AND RowTime >= 1510923225000
+        .Select(l => new { l.Id, l.Message })
         .Take(2)     
         .ToObservable() // client side processing starts here lazily after subscription
         .ObserveOn(TaskPoolScheduler.Default)
-
         .Subscribe(tweetMessage =>
         {
           Console.WriteLine($"{nameof(Tweet)}: {tweetMessage.Id} - {tweetMessage.Message}");
+          Console.WriteLine();
         }, error => { Console.WriteLine($"Exception: {error.Message}"); }, () => Console.WriteLine("Completed"));
 
       Console.WriteLine("Press any key to stop the subscription");
@@ -44,6 +46,7 @@ namespace Kafka.DotNet.ksqlDB.Sample
 
       return subscriptions;
     }
+
     private static IDisposable ToObservableExample(string ksqlDbUrl)
     {
       var subscriptions = new KQueryStreamSet<Tweet>(new QbservableProvider(ksqlDbUrl))
@@ -52,6 +55,14 @@ namespace Kafka.DotNet.ksqlDB.Sample
         .Subscribe(new TweetsObserver());
 
       return subscriptions;
+    }
+
+    private static void ToQueryStringExample(string ksqlDbUrl)
+    {
+      var ksql = new TweetsQueryStream(ksqlDbUrl).ToQueryString();
+
+      //prints SELECT * FROM s EMIT CHANGES;
+      Console.WriteLine(ksql);
     }
   }
 }