+++ Tests/Joker.Kafka.Tests/Extensions/KSql/RestApi/TestableKSqlDbQueryStreamProvider.cs
@@ -21,12 +21,12 @@ namespace Kafka.DotNet.ksqlDB.Tests.Extensions.KSql.RestApi
     protected override HttpClient OnCreateHttpClient()
     {      
       var queryResponse =
-        @"{""queryId"":""59df818e-7d88-436f-95ac-3c59becc9bfb"",""columnNames"":[""MESSAGE"",""ID"",""ISROBOT"",""AMOUNT""],""columnTypes"":[""STRING"",""INTEGER"",""BOOLEAN"",""DOUBLE""]}
-[""Hello world"",1,true,0.1]
-[""Wall-e"",2,false,1.1]";
+        @"{""queryId"":""59df818e-7d88-436f-95ac-3c59becc9bfb"",""columnNames"":[""ROWTIME"",""MESSAGE"",""ID"",""ISROBOT"",""AMOUNT""],""columnTypes"":[""BIGINT"",""STRING"",""INTEGER"",""BOOLEAN"",""DOUBLE""]}
+[1611327570881,""Hello world"",1,true,0.1]
+[1611327671476,""Wall-e"",2,false,1.1]";
 
       var errorResponse =
-        @"{""@type"":""generic_error"",""error_code"":40001,""message"":""Line: 1, Col: 21: SELECT column 'KVAK' cannot be resolved.\nStatement: SELECT Message, Id, Kvak FROM Tweets\r\nWHERE Message = 'Hello world' EMIT CHANGES LIMIT 2;""}";
+        @"{""@type"":""generic_error"",""error_code"":40001,""message"":""Line: 1, Col: 21: SELECT column 'Foo' cannot be resolved.\nStatement: SELECT Message, Id, Foo FROM Tweets\r\nWHERE Message = 'Hello world' EMIT CHANGES LIMIT 2;""}";
 
       var handlerMock = new Mock<HttpMessageHandler>();